+++ CHANGELOG.md
@@ -9,6 +9,13 @@ that bundle.
 
 Conventions:
 
+## 4.3.52
+
+Release Date: 2026-08-02
+
+### Added
+- Refactored rules deployer script using modern v14 modular app and security-rules APIs.
+
 ## 4.3.51
 
 Release Date: 2026-08-02